update cli exec metadata creation
